feat: reimplement video
Reintroduces the video player, which is now contained in the outer
`Layout` component, allowing for reusing the player across pages and,
thus, removing the need to get unmute permission repeatedly.

The video player is controlled via `LayoutContext.video`. Both `Banner`
and `MainContent` observe some properties of `video` and change their
layout accordingly.

The video to be played is defined in `customData` and enabled for
questions and question categories. Coverage should be added to other
pages as well, see #723

Also:

- Centralize `customData` parsing in `StrapiDataProvider`
- Add `video` to `CustomData.QuestionCategory`
- Move the `main` element from `MainContent` to `Layout` so that the
  video player is contained within it